Polish BasicTile to use realistic layout
Modeled similar to the BasicTile component we use in our template app.

Thumbnail on the left side, with vertical list of textviews for title,
subtilte, and description on right.

Moved the clicks text under the thumbnail and made it's fontweight
dynamic based on whether there are more than 0 clicks.

Add Constants file for using custom constant strings and values.

Also add images folder for local images, which we currently use
for the BasicTile thumbnail